[change] ディレクトリ構造を変更
