+++ app/javascript/app/redux/configure-store.jsx
@@ -2,10 +2,11 @@ import 'babel-polyfill';
 import { createStore, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 
-import reducer, { rootSaga } from './modules/reducer';
+import rootReducer from './modules/reducer';
+import rootSaga from './modules/sagas';
 
 const sagaMiddleware = createSagaMiddleware();
-const store = createStore(reducer, applyMiddleware(sagaMiddleware));
+const store = createStore(rootReducer, applyMiddleware(sagaMiddleware));
 
 sagaMiddleware.run(rootSaga);
