+++ app/javascript/app/redux/modules/emojis/actions.jsx
@@ -1,0 +1,9 @@
+import * as types from './types';
+
+export function loadEmojis() {
+  return { type: types.LOAD };
+}
+
+export function sucessLoadEmojis(emojis) {
+  return { type: types.SUCCESS_LOAD, emojis };
+}
