+++ app/javascript/app/redux/modules/emojis/types.jsx
@@ -1,0 +1,2 @@
+export const LOAD = 'emojis/LOAD';
+export const SUCCESS_LOAD = 'emojis/SUCCESS_LOAD';
